refactor: DTO 객체를 별도 패키지로 분리
