+++ src/main/java/roomescape/controller/ReservationController.java
@@ -1,5 +1,6 @@
 package roomescape.controller;
 
+import java.util.List;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -8,10 +9,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
+import roomescape.dto.ReservationReqDto;
 import roomescape.model.Reservation;
-import roomescape.model.ReservationReqDto;
 import roomescape.model.Reservations;
 
 @RestController
