+++ src/main/java/roomescape/dto/ReservationReqDto.java
@@ -1,4 +1,4 @@
-package roomescape.model;
+package roomescape.dto;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
