+++ src/test/java/roomescape/ReservationsTest.java
@@ -4,8 +4,8 @@ import java.time.LocalDate;
 import java.time.LocalTime;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import roomescape.dto.ReservationReqDto;
 import roomescape.model.Reservation;
-import roomescape.model.ReservationReqDto;
 import roomescape.model.Reservations;
 
 import static org.assertj.core.api.Assertions.assertThat;
